Lab 4
In this commit, I have changed the story of my game which has been changed in the README file. The mechanics are still very similar and the Health bar reaches the requirements for Lab 4. I will continue to build the village and the world around during reading week. Please read the README file for a short explanation of the game.